Reddit-Topic-Explorer/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
bertopic==0.17.4
blinker==1.9.0
certifi==2025.11.12
cffi 
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
cssselect2 
distro==1.9.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fonttools 
fsspec==2025.12.0
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
kaleido==1.2.0
llvmlite==0.46.0
logistro==2.0.1
MarkupSafe==3.0.3
mpmath==1.3.0
munkres==1.1.4
narwhals==2.13.0
networkx==3.6.1
numba==0.63.0
numpy==2.3.5
openai==2.9.0
orjson==3.11.5
packaging 
pandas==2.3.3
pillow 
plotly==6.5.0
pluggy==1.6.0
pycparser 
pydantic==2.12.5
pydantic_core==2.41.5
pydyf 
Pygments==2.19.2
pynndescent==0.5.13
pyphen 
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2 
tinyhtml5 
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
unicodedata2 
urllib3==2.6.1
vaderSentiment==3.3.2
weasyprint 
webencodings
Werkzeug==3.1.4
